feat: update vim configuration
I have used two shell scripts to maintain the vim plugins. However, it's
a bad idea to use raw shell scripts to do this. Instead, I choose a vim
plugin "vim-plug" to do this and install the plugins.

Because I only use `vim` to do some basic operations, thus I reduce the
plugin of the coc.nvim. However, we need to disable "sensible" of
"vim-polyglot" in order to make sure that indentation is not changed.

Last, update some keys to align with VsCode/Cursor configuration and
delete the shell scripts.