archlinux: write script to control volume
Because we have deleted pa-applet, we cannot use keyboard to control the
volume, we need to add the keybingds in the i3 config.

In order to do that create a shell script "volume_control.sh" to do that
and use `dunstify` to notify the user and in order not to create the
history, update the `dunstrc` to ignore app name "i3_volume_control".