archlinux(polybar): fix some issues
We misuse fontawesome(fa) and material design(md). Let's use the format
of material design.

And we use circle icons which is hard to know which workspace we are at.
So, we should use icon with number. However, the icon size is small.
Let's introduce a new font with bigger size for i3 module.

Also, there is no need to add icon for date. Let's delete the icon. And
add the forgotten "spotify.ini" module configuration.